moved rtc_irqhandler into time.cpp
Fixes #1
